+++ elfo/tests/streams.rs
@@ -1,6 +1,8 @@
 #![cfg(feature = "full")]
 
-use elfo::{config::AnyConfig, prelude::*, stream};
+use std::convert::TryFrom;
+
+use elfo::{_priv::tls, config::AnyConfig, prelude::*, stream, trace_id::TraceId};
 
 #[message]
 #[derive(PartialEq)]
@@ -12,18 +14,18 @@ struct Set(Vec<u32>);
 #[message]
 struct Replace(Vec<u32>);
 
-#[message]
-#[derive(PartialEq)]
-struct EndOfMessages;
-
 #[tokio::test]
 async fn it_handles_basic_operations() {
     let group = ActorGroup::new().exec(|ctx| async move {
         let stream = stream::Stream::new(futures::stream::iter(vec![SomeMessage(0)]));
 
         let mut ctx = ctx.with(&stream);
+        let mut prev_trace_id = tls::trace_id();
 
         while let Some(envelope) = ctx.recv().await {
+            assert_ne!(tls::trace_id(), prev_trace_id);
+            prev_trace_id = tls::trace_id();
+
             msg!(match envelope {
                 m @ SomeMessage(_) => ctx.send(m).await.unwrap(),
                 Set(d) => stream.set(futures::stream::iter(
@@ -34,7 +36,6 @@ async fn it_handles_basic_operations() {
                         d.into_iter().map(SomeMessage).collect::<Vec<_>>(),
                     ));
                 }
-                EndOfMessages => ctx.send(EndOfMessages).await.unwrap(),
             })
         }
     });
@@ -53,4 +54,29 @@ async fn it_handles_basic_operations() {
     for i in 6..8 {
         assert_msg_eq!(proxy.recv().await, SomeMessage(i));
     }
+}
+
+#[tokio::test]
+async fn it_restores_trace_id() {
+    let group = ActorGroup::new().exec(|ctx| async move {
+        let stream = stream::Stream::new(futures::stream::iter(vec![
+            (TraceId::try_from(5).unwrap(), SomeMessage(5)),
+            (TraceId::try_from(6).unwrap(), SomeMessage(6)),
+        ]));
+
+        let mut ctx = ctx.with(&stream);
+
+        while let Some(envelope) = ctx.recv().await {
+            msg!(match envelope {
+                SomeMessage(x) => {
+                    assert_eq!(u64::from(tls::trace_id()), u64::from(x));
+                    ctx.send(SomeMessage(x)).await.unwrap()
+                }
+            })
+        }
+    });
+
+    let mut proxy = elfo::test::proxy(group, AnyConfig::default()).await;
+    assert_msg_eq!(proxy.recv().await, SomeMessage(5));
+    assert_msg_eq!(proxy.recv().await, SomeMessage(6));
 }